# src/type/request/Queue.request.ts
// export interface SaveReqOrder{
//     Customer_name:string;
//     Table_no:string;
//     Waiter:object;
//     Food:object[];
//     Order_status:number;
//     }
    export interface UpdateReqOrderQueue {
      _id:string;
    
      
    }
    // export interface GetOrderQueue {
    //   id: string
    // }
    export interface deleteorderQueue {
      id: string
    }
    export interface SearchReqOrder{
      product:object
    }